requirements.txt
aioredis==2.0.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.0
asyncio==3.4.3
asyncpg==0.30.0
asyncpgsa==0.27.1
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
fastapi==0.115.4
greenlet==3.1.1
h11==0.14.0
idna==3.10
pyjwt
Mako==1.3.6
MarkupSafe==3.0.2
numpy==2.1.3
pandas==2.2.3
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.0
